Shark encounter
The shark encounter is similar to the seagull encounter you are
given the option to scare if off if your random number roll is higher than the set one which is > 8.
You are also given the option to feed the shark which will make it happy, to play with it where it will be happy and go away,
and a null help function which does nothing.