ONE BIG COMMIT
THe island has two areas one mountain and one jungle area.
The jungle area has a boss entitled the Monkey King which is in a sub-location
that you can only get to once your in the jungle. During the boss battle when
the boss reaches a certain health his attacks update so he has a new attack.
when defeated he provides info to help with the puzzle in the mountain area.
In the mountain area there is an altar that when you say yes to touching it
sends you to the riddle sub location where the puzzles are.
There are 3 puzzles ranging in difficulty providing ship food, food and
coin that provides overall points, and a weapon respectively.
If you guess a question wrong you can access this area but will be transported
away the second that you move. The mountain area has a goat monster as well which
is similar to the Macaque. Each area has 4 ways of motion that prompt text.