Abandoned Island Version 1.0 Release
The entire Island is now complete! At least for the initial release.
There are currently no plans for future updates unless there is a game-breaking crash that needs to be fixed.

The island in its current state provides a good adventure for anyone who stumbles across it, despite its relatively small size.
There are a number of things that can be found around this island, some of which are mentioned below.

Features of this release include:
 - The chance for a thrilling seagull attack at every beach you walk on.
 - A puzzle encompassing the entire island.
 - A few special loot items.
 - A Knights and Knaves puzzle.
 - A weather-beaten church with some interesting things inside...
 - And anything else I forgot to mention!